+++ cmd/dagger/listen.go
@@ -35,7 +35,7 @@ func init() {
 func Listen(cmd *cobra.Command, args []string) {
 	ctx := context.Background()
 	if err := withEngineAndTUI(ctx, client.Params{SecretToken: os.Getenv("DAGGER_SESSION_TOKEN")}, func(ctx context.Context, engineClient *client.Client) error {
-		rec := progrock.RecorderFromContext(ctx)
+		rec := progrock.FromContext(ctx)
 
 		var stderr io.Writer
 		if silent {
