+++ core/directory.go
@@ -599,7 +599,7 @@ func (dir *Directory) Export(
 		defPB = dir.LLB
 	}
 
-	rec := progrock.RecorderFromContext(ctx)
+	rec := progrock.FromContext(ctx)
 
 	vtx := rec.Vertex(
 		digest.Digest(identity.NewID()),
